+++ internal/abi/BidderRegistry.abi
