+++ internal/abi/ValidatorRegistry.abi
